第二次提交（没头绪）

+++ XueJi_Manage/src/main/java/com/fc/dao/TAdminMapper.java
@@ -21,7 +21,7 @@ public interface TAdminMapper {
 
     List<TAdmin> selectByExample(TAdminExample example);
 
-    TAdmin selectByPrimaryKey(Integer userid);
+    TAdmin selectByPrimaryKey(Long userid);
 
     int updateByExampleSelective(@Param("record") TAdmin record, @Param("example") TAdminExample example);
